added JSON-based formatting and labelling options, dropped writing index from output
Formatting options are something of a preprocessing option to rework the output of columns prior to partitioning the data into separate files.  Essentially to enforce consistency in the partitioning values.

The labels option is to allow for the option of adding column names to the output files for specifically matching partition values/files.