added fix for unconventional times and drop columns for csv slicer
Some valid ISO 8601 times can use 24:00 to represent 00:00 - while valid it is not recognized by strptime().  When an error parsing datetimes occurs, the offending string and parsing format are displayed, then an attempt will be made to string replace 24:00 or 2400 with 00:00 or 0000 respectively.  These seem like the most probable forms that problematic times would be represented as and parsing will be attempted again.  There is no secondary catch if this fails.

The --drop-columns option has been ported to csv_slicer.py from csv_convert_date.py